Refactors app to new class structure
