+++ routes/messages/commands/get_books_from_category.js
@@ -1,14 +1,17 @@
-const Categories = require('../../../models/db/categories');
+const Categories = require('models/db/categories');
 const getBooksInCategories = require('../helpers/getBooksInCategories.js');
-const BookTemplate = require('../UI/BookTemplate.js');
-const QuickReplyTemplate = require('../UI/QuickReplyTemplate.js');
+const BookTemplate = require('../Templates/Book.js');
+const QuickReplyTemplate = require('../Templates/QuickReply.js');
 const browse = require('./browse.js');
 
-module.exports = async (event) => {
+module.exports = async event => {
   const { category_id, user_id } = event;
   const currentCat = await Categories.retrieve({ id: category_id });
 
-  const { isEndOfCategory, books } = await getBooksInCategories(user_id, category_id);
+  const { isEndOfCategory, books } = await getBooksInCategories(
+    user_id,
+    category_id
+  );
 
   const text = `Would you like to see more books on ${currentCat[0].name}?`;
   const quickReplies = [];
@@ -24,19 +27,22 @@ module.exports = async (event) => {
     }
   ];
 
-  options.forEach((opt) => {
+  options.forEach(opt => {
     const { title, command, category_id } = opt;
 
     quickReplies.push({
       title,
-      payload: JSON.stringify({ 
+      payload: JSON.stringify({
         command: command.toLowerCase(),
-        category_id 
+        category_id
       })
     });
   });
   const browseQR = await browse(event);
   return isEndOfCategory
-    ? [ await BookTemplate(event, books), ...browseQR ]
-    : [ await BookTemplate(event, books), await QuickReplyTemplate(text, quickReplies) ];
+    ? [await BookTemplate(event, books), ...browseQR]
+    : [
+        await BookTemplate(event, books),
+        await QuickReplyTemplate(text, quickReplies)
+      ];
 };
