+++ routes/messages/commands/start_premium.js
@@ -1,4 +1,4 @@
-const SubscribeTemplate = require('../UI/SubscribeTemplate.js');
+const SubscribeTemplate = require('../Templates/Subscribe.js');
 
 module.exports = async event => {
   const { user_id } = event;
